refactor: refactor search-bar into user-mfe + add federated-component for displaying the new user-mfe (search-bar)
