+++ client/projects/shell/src/app/app.component.html
@@ -30,22 +30,10 @@
 </mat-drawer-container>
 
 <ng-template #searchInput>
-  <form>
-    <mat-form-field appearance="fill" style="font-size: 12px; height: 2.3em">
-      <mat-label>Search</mat-label>
-      <input type="text"
-             placeholder="Username"
-             matInput
-             [formControl]="searchFormControl"
-             [matAutocomplete]="auto">
-      <button matSuffix mat-icon-button aria-label="Clear" (click)="closeSearch()" type="reset">
-        <mat-icon>close</mat-icon>
-      </button>
-      <mat-autocomplete #auto="matAutocomplete">
-        <mat-option *ngFor="let option of filteredOptions | async" [value]="option" (click)="redirect(option)">
-          {{option}}
-        </mat-option>
-      </mat-autocomplete>
-    </mat-form-field>
-  </form>
+  <div class="search-container">
+    <app-federated-component remoteEntry="http://localhost:5003/remoteEntry.js" remoteName="user" exposedModule="UserModule" componentName="UserComponent"></app-federated-component>
+    <button mat-icon-button (click)="closeSearch()">
+      <mat-icon>search_off</mat-icon>
+    </button>
+  </div>
 </ng-template>
